Moved logging and config services to a common directory.  Added global error handler.
