+++ webapp/src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { AppConfig } from './config/app.config';
+import { AppConfig } from './common/config/app.config';
 import { VERSION } from 'src/environments/version';
 import { Angulartics2GoogleAnalytics } from 'angulartics2/ga';
