+++ webapp/src/app/app.module.ts
@@ -1,9 +1,10 @@
-import { APP_INITIALIZER, NgModule } from '@angular/core';
+import { APP_INITIALIZER, ErrorHandler, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { Angulartics2Module } from 'angulartics2';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { AppConfig } from './config/app.config';
+import { AppConfig } from './common/config/app.config';
+import { GlobalErrorHandler } from './common/global-error-handler';
 import { DataModule } from './data/data.module';
 import { HomeModule } from './home/home.module';
 import { LayoutModule } from './layout/layout.module';
@@ -21,6 +22,7 @@ export function initializeApp(appConfig: AppConfig) {
     Angulartics2Module.forRoot(),
     BrowserModule,
     AppRoutingModule,
+    
     DataModule,
     LayoutModule,
     HomeModule,
@@ -29,8 +31,11 @@ export function initializeApp(appConfig: AppConfig) {
   providers: [{ 
     provide: APP_INITIALIZER,
     useFactory: initializeApp,
-    deps: [ AppConfig ], multi: true }
-  ],
+    deps: [ AppConfig ], multi: true 
+  }, {
+    provide: ErrorHandler,
+    useValue: GlobalErrorHandler
+  }],
   bootstrap: [ AppComponent ]
 })
 export class AppModule { }
