+++ webapp/src/app/common/logging/logging.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 import { LoggingService } from './logging.service';
-import { initializeSettingsProvider } from '../data/mock-data';
+import { initializeSettingsProvider } from '../../data/mock-data';
 
 describe('LoggingService', () => {
   beforeEach(() => TestBed.configureTestingModule({
